requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
ta==0.10.2
python-dateutil==2.8.2
matplotlib==3.8.0
mplfinance==0.12.9b7
